Condense baseProcessors with utils
